chore: add test for the name duplication when creating stories. Also did some small refactoring.
